+++ workspace/TripTrack/inc/dao/TTLocation.h
@@ -23,7 +23,7 @@ public:
 	TTLocation();
 	virtual ~TTLocation();
 
-	result Construct(int id);
+	result Construct(long long int id);
 	result Construct(Tizen::Locations::Location location);
 	Tizen::Locations::Coordinates* getCoordinates(void);
 	Tizen::Base::DateTime* getTimestamp(void);
@@ -40,14 +40,14 @@ public:
 	virtual Tizen::Io::DbStatement* Write(void);
 	virtual Tizen::Io::DbStatement* Delete(void);
 	virtual Tizen::Io::DbStatement* Update(void);
-	int GetLocationId() const;
+	long long int GetLocationId() const;
 
 private:
 	Tizen::Locations::Coordinates* __pCoordinates;
 	Tizen::Base::DateTime* __pTimeStamp;
 	double __speed;
 	double __course;
-	int __locationId;
+	long long int __locationId;
 };
 
 #endif /* TTLOCATION_H_ */
